+++ src/components/common/Alarm.tsx
@@ -1,16 +1,12 @@
 import { useEffect, useState } from 'react';
-// 라이브러리
 import { useQuery } from '@tanstack/react-query';
-import { ToastContainer, toast } from 'react-toastify';
+import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-// api
+
 import { supabase } from '../../api/supabase';
 import { getSubList } from '../../api/subscribe';
-// zustand 상태관리 hook
 import { useCurrentUser } from '../../store/userStore';
-// 스타일
-import { styled } from 'styled-components';
-// mui
+
 import TextsmsIcon from '@mui/icons-material/Textsms';
 import BorderColorIcon from '@mui/icons-material/BorderColor';
 import PeopleAltIcon from '@mui/icons-material/PeopleAlt';
@@ -206,53 +202,7 @@ const Alarm = () => {
     }
   }, [msgData, currentUserId]);
 
-  return (
-    <div>
-      {/* <AlarmContainer
-        position="top-right"
-        autoClose={3000}
-        // hideProgressBar={true}
-        newestOnTop={true}
-        // closeOnClick={true}
-        // rtl={true}
-        pauseOnFocusLoss={false}
-        draggable={true}
-        pauseOnHover={true}
-        // limit={3}
-      /> */}
-    </div>
-  );
+  return <></>;
 };
 
 export default Alarm;
-
-export const AlarmContainer = styled(ToastContainer)`
-  .custom-toast {
-    background-color: var(--sixth-color);
-    color: black;
-  }
-  .Toastify__toast {
-    font-size: 15px;
-    /* border-radius: 50px; */
-    padding: 15px 20px;
-    color: #ffffff;
-    /* background: rgba(107, 115, 135, 0.8); */
-  }
-
-  .Toastify__toast-icon {
-    width: 20px;
-    height: 20px;
-  }
-
-  .Toastify__toast--info {
-    background: #136e65;
-  }
-
-  /* .Toastify__toast--success {
-    background: rgba(48, 173, 120, 0.8);
-  } */
-
-  /* .Toastify__toast--error {
-    background: rgba(224, 72, 82, 0.8);
-  } */
-`;
